+++ src/App.js
@@ -1,21 +1,20 @@
-import React, { PureComponent } from "react";
+import React, { PureComponent, Fragment } from "react";
 import { Route, Switch } from "react-router-dom";
 
 //Components
 import Gallery from "./Components/gallery";
 import Landing from "./Components/landing";
-import NoMatch from "./Components/no-match-page";
+//import NoMatch from "./Components/no-match-page"; <Route component={NoMatch} />
 
 class App extends PureComponent {
   render() {
     return (
-      <React.Fragment>
-        <Landing />
+      <Fragment>
         <Switch>
+          <Route exact path="/" component={Landing} /> 
           <Route path="/gallery" component={Gallery} />
-          <Route component={NoMatch} />
         </Switch>
-      </React.Fragment>
+      </Fragment>
     );
   }
 }